omsBackend/requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
certifi==2017.11.5
chardet==3.0.4
Django==1.11.4
django-channels==0.7.0
django-cors-headers==2.1.0
django-filter==1.0.4
django-rest-auth==0.9.1
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
dry-rest-permissions==0.1.9
idna==2.6
msgpack-python==0.4.8
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
PyJWT==1.5.3
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22